move roomsystem and shaderstruct back since non binary
