+++ CompilerSource/compiler/components/write_shader_data.cpp
@@ -58,7 +58,7 @@ int lang_CPP::compile_writeShaderData(const GameData &game, parsed_object *EGMgl
 {
   ofstream wto((codegen_directory + "Preprocessor_Environment_Editable/IDE_EDIT_shaderarrays.h").c_str(),ios_base::out);
 
-  wto << license << "#include \"Universal_System/Resources/shaderstruct.h\"\n" << "namespace enigma {\n";
+  wto << license << "#include \"Universal_System/shaderstruct.h\"\n" << "namespace enigma {\n";
   wto << "  std::vector<ShaderStruct> shaderstructarray = {\n";
 
   int idmax = 0;
