+++ ENIGMAsystem/SHELL/Collision_Systems/BBox/BBOXfuncs.cpp
@@ -21,7 +21,7 @@
 
 #include "Universal_System/Object_Tiers/collisions_object.h"
 #include "Universal_System/Instances/instance_system.h" //iter
-#include "Universal_System/Resources/roomsystem.h"
+#include "Universal_System/roomsystem.h"
 #include "Collision_Systems/collision_mandatory.h" //iter
 #include "BBOXimpl.h"
 #include "../General/CSfuncs.h"
