+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D11/DIRECTX11Std.cpp
@@ -22,7 +22,7 @@
 #include "Graphics_Systems/graphics_mandatory.h" // Room dimensions.
 
 #include "Universal_System/var4.h"
-#include "Universal_System/Resources/roomsystem.h" // Room dimensions.
+#include "Universal_System/roomsystem.h" // Room dimensions.
 
 #include <iostream>
 #include <string>
