+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D11/DX11screen.cpp
@@ -26,7 +26,7 @@
 #include "Universal_System/nlpo2.h"
 #include "Universal_System/Resources/background_internal.h"
 #include "Universal_System/Resources/sprites_internal.h"
-#include "Universal_System/Resources/roomsystem.h"
+#include "Universal_System/roomsystem.h"
 #include "Platforms/General/PFwindow.h"
 
 #include <string>
