+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D9/DIRECTX9Std.cpp
@@ -19,7 +19,7 @@
 
 #include "DIRECTX9Std.h"
 #include "Universal_System/var4.h"
-#include "Universal_System/Resources/roomsystem.h" // Room dimensions.
+#include "Universal_System/roomsystem.h" // Room dimensions.
 #include "Graphics_Systems/graphics_mandatory.h" // Room dimensions.
 
 #include <iostream>
