+++ ENIGMAsystem/SHELL/Graphics_Systems/None/fillin.cpp
@@ -29,10 +29,10 @@
 #include "Graphics_Systems/General/GSmatrix.h"
 #include "Graphics_Systems/General/GScolors.h"
 #include "Collision_Systems/collision_types.h"
-#include "Universal_System/Resources/shaderstruct.h"
+#include "Universal_System/shaderstruct.h"
 #include "Universal_System/scalar.h"
 #include "Universal_System/var4.h"
-#include "Universal_System/Resources/roomsystem.h" // Room dimensions.
+#include "Universal_System/roomsystem.h" // Room dimensions.
 #include "Universal_System/depth_draw.h"
 #include "Universal_System/Resources/background_internal.h"
 #include "Universal_System/Resources/sprites_internal.h"
