+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GL3d3d.cpp
@@ -32,7 +32,7 @@
 #include "Graphics_Systems/General/GScolor_macros.h"
 
 #include "Widget_Systems/widgets_mandatory.h"
-#include "Universal_System/Resources/roomsystem.h" // for view variables
+#include "Universal_System/roomsystem.h" // for view variables
 
 #include <glm/gtc/type_ptr.hpp>
