+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GL3screen.cpp
@@ -32,7 +32,7 @@
 #include "Universal_System/nlpo2.h"
 #include "Universal_System/Resources/background_internal.h"
 #include "Universal_System/Resources/sprites_internal.h"
-#include "Universal_System/Resources/roomsystem.h"
+#include "Universal_System/roomsystem.h"
 #include "Platforms/General/PFwindow.h"
 
 #include <string>
