+++ ENIGMAsystem/SHELL/Platforms/Win32/WINDOWSmain.cpp
@@ -25,7 +25,7 @@
 
 #include "Universal_System/mathnc.h" // enigma_user::clamp
 #include "Universal_System/estring.h"
-#include "Universal_System/Resources/roomsystem.h"
+#include "Universal_System/roomsystem.h"
 #include "Universal_System/var4.h"
 
 #include <mmsystem.h>
