+++ ENIGMAsystem/SHELL/Platforms/iPhone/CocoaFunctions.cpp
@@ -29,7 +29,7 @@
 #include "CocoaWindow.h"
 #include <stdio.h>
 #include "Platforms/General/PFmain.h"
-#include "Universal_System/Resources/roomsystem.h"
+#include "Universal_System/roomsystem.h"
 
 
 namespace enigma
