+++ ENIGMAsystem/SHELL/Platforms/iPhone/CocoaMain.cpp
@@ -23,7 +23,7 @@
 #include "ObjectiveC.h"
 
 #include "Platforms/General/PFmain.h"
-#include "Universal_System/Resources/roomsystem.h"
+#include "Universal_System/roomsystem.h"
 
 #include <stdio.h>
 #include <iostream>
