+++ ENIGMAsystem/SHELL/Universal_System/Extensions/GM5Compat/GM5Views.cpp
@@ -19,7 +19,7 @@
 
 #include "GM5Views.h"
 
-#include "Universal_System/Resources/roomsystem.h"
+#include "Universal_System/roomsystem.h"
 
 
 namespace enigma_user
