+++ ENIGMAsystem/SHELL/Universal_System/depth_draw.h
@@ -36,7 +36,7 @@
 #define ENIGMA_DEPTH_DRAW_H
 
 #include "Instances/instance_system.h"
-#include "Resources/roomsystem.h"
+#include "roomsystem.h"
 
 #include <map>
 #include <set>
