+++ ENIGMAsystem/SHELL/Universal_System/shaderstruct.cpp
